pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "oltp-warehouse"
version = "0.1.0"
description = "Synthetic OLTP bootstrap generator for a warehouse pipeline project."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.10,<1.11",
    "dbt-duckdb>=1.9,<1.10",
    "pyarrow>=20.0,<21.0",
    "psycopg[binary]>=3.2,<4.0",
]

[project.scripts]
oltp-warehouse = "oltp_warehouse.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
